+++ Models/Items/Item.cs
@@ -8,7 +8,7 @@ public abstract class Item
     public string Name { get; protected set; }
     public string Ean { get; protected set; }
     public int Quantity { get; protected set; }
-    public float Price { get; protected set; }
+    public double Price { get; protected set; }
     public string ImageUrl { get; protected set; }
     public string Description { get; protected set; }
     public float Mass { get; protected set; }
